+++ localstack/services/cloudformation/provider.py
@@ -4,6 +4,7 @@ import re
 from collections import defaultdict
 from copy import deepcopy
 
+import localstack.services.cloudformation.api_utils
 from localstack.aws.api import CommonServiceException, RequestContext, handler
 from localstack.aws.api.cloudformation import (
     CallAs,
@@ -130,7 +131,9 @@ class CloudformationProvider(CloudformationApi):
     @handler("CreateStack", expand=False)
     def create_stack(self, context: RequestContext, request: CreateStackInput) -> CreateStackOutput:
         state = get_cloudformation_store()
-        template_preparer.prepare_template_body(request)  # TODO: avoid mutating request directly
+        localstack.services.cloudformation.api_utils.prepare_template_body(
+            request
+        )  # TODO: avoid mutating request directly
         template = template_preparer.parse_template(request["TemplateBody"])
         stack_name = template["StackName"] = request.get("StackName")
         stack = Stack(request, template)  # TODO: proper body handling like in create_change_set
@@ -186,7 +189,7 @@ class CloudformationProvider(CloudformationApi):
         if not stack:
             return not_found_error(f'Unable to update non-existing stack "{stack_name}"')
 
-        template_preparer.prepare_template_body(request)
+        localstack.services.cloudformation.api_utils.prepare_template_body(request)
         template = template_preparer.parse_template(request["TemplateBody"])
         new_stack = Stack(request, template)
         deployer = template_deployer.TemplateDeployer(stack)
@@ -283,7 +286,7 @@ class CloudformationProvider(CloudformationApi):
             if not stack:
                 return stack_not_found_error(stack_name)
         else:
-            template_preparer.prepare_template_body(request)
+            localstack.services.cloudformation.api_utils.prepare_template_body(request)
             template = template_preparer.parse_template(request["TemplateBody"])
             request["StackName"] = "tmp-stack"
             stack = Stack(request, template)
@@ -330,7 +333,7 @@ class CloudformationProvider(CloudformationApi):
                 "Specify exactly one of 'TemplateBody' or 'TemplateUrl'"
             )  # TODO: check proper message
 
-        template_preparer.prepare_template_body(
+        localstack.services.cloudformation.api_utils.prepare_template_body(
             req_params
         )  # TODO: function has too many unclear responsibilities
         if not template_body:
@@ -600,7 +603,7 @@ class CloudformationProvider(CloudformationApi):
     ) -> ValidateTemplateOutput:
         try:
             # TODO implement actual validation logic
-            template_body = template_preparer.get_template_body(request)
+            template_body = localstack.services.cloudformation.api_utils.get_template_body(request)
             valid_template = json.loads(template_preparer.template_to_json(template_body))
 
             parameters = [
