+++ tests/unit/test_cloudformation.py
@@ -1,11 +1,12 @@
 import re
 from typing import Dict
 
+import localstack.services.cloudformation.api_utils
 from localstack.services.cloudformation.deployment_utils import (
     PLACEHOLDER_AWS_NO_VALUE,
     remove_none_values,
 )
-from localstack.services.cloudformation.engine import template_deployer, template_preparer
+from localstack.services.cloudformation.engine import template_deployer
 from localstack.services.cloudformation.engine.entities import Stack
 from localstack.services.cloudformation.models.stepfunctions import _apply_substitutions
 
@@ -52,9 +53,9 @@ def test_is_local_service_url():
         "http://mybucket.s3.us-east-1.amazonaws.com",
     ]
     for url in local_urls:
-        assert template_preparer.is_local_service_url(url)
+        assert localstack.services.cloudformation.api_utils.is_local_service_url(url)
     for url in remote_urls:
-        assert not template_preparer.is_local_service_url(url)
+        assert not localstack.services.cloudformation.api_utils.is_local_service_url(url)
 
 
 def test_apply_substitutions():
